Group local date related tests
